src/server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0

sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1

pydantic-settings==2.2.1

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

python-multipart==0.0.9

aio-pika==9.4.1

pyyaml==6.0.1

httpx==0.27.0

pytest==8.2.0
pytest-cov==7.1.0

loguru==0.7.2